Clean up comments and extra files
