+++ cmd/export_operations.go
@@ -22,7 +22,7 @@ var operationsCmd = &cobra.Command{
 		startNum, path, limit := utils.MustArchiveFlags(cmd.Flags(), cmdLogger)
 		env := utils.GetEnvironmentDetails(isTest)
 
-		operations, err := input.GetOperationsCaptive(startNum, endNum, limit, env)
+		operations, err := input.GetOperations(startNum, endNum, limit, env)
 		if err != nil {
 			cmdLogger.Fatal("could not read operations: ", err)
 		}
