+++ internal/input/operations.go
@@ -2,12 +2,13 @@ package input
 
 import (
 	"context"
+	"fmt"
 	"io"
 
-	"github.com/stellar/stellar-etl/internal/utils"
-
 	"github.com/stellar/go/ingest"
+	"github.com/stellar/go/ingest/ledgerbackend"
 	"github.com/stellar/go/xdr"
+	"github.com/stellar/stellar-etl/internal/utils"
 )
 
 // OperationTransformInput is a representation of the input for the TransformOperation function
@@ -18,21 +19,46 @@ type OperationTransformInput struct {
 	LedgerSeqNum   int32
 }
 
+func panicIf(err error) {
+	if err != nil {
+		panic(fmt.Errorf("An error occurred, panicking: %s\n", err))
+	}
+}
+
 // GetOperations returns a slice of operations for the ledgers in the provided range (inclusive on both ends)
-func GetOperations(start, end uint32, limit int64, isTest bool) ([]OperationTransformInput, error) {
-	env := utils.GetEnvironmentDetails(isTest)
-	backend, err := utils.CreateBackend(start, end, env.ArchiveURLs)
+func GetOperations(start, end uint32, limit int64, env utils.EnvironmentDetails) ([]OperationTransformInput, error) {
+	ctx := context.Background()
+	captiveCoreToml, err := ledgerbackend.NewCaptiveCoreTomlFromFile(
+		env.CoreConfig,
+		ledgerbackend.CaptiveCoreTomlParams{
+			NetworkPassphrase:  env.NetworkPassphrase,
+			HistoryArchiveURLs: env.ArchiveURLs,
+			Strict:             true,
+		},
+	)
+
+	backend, err := ledgerbackend.NewCaptive(
+		ledgerbackend.CaptiveCoreConfig{
+			BinaryPath:         env.BinaryPath,
+			Toml:               captiveCoreToml,
+			NetworkPassphrase:  env.NetworkPassphrase,
+			HistoryArchiveURLs: env.ArchiveURLs,
+		},
+	)
 	if err != nil {
 		return []OperationTransformInput{}, err
 	}
 
 	opSlice := []OperationTransformInput{}
-	ctx := context.Background()
+	err = backend.PrepareRange(ctx, ledgerbackend.BoundedRange(start, end))
+	panicIf(err)
 	for seq := start; seq <= end; seq++ {
-		txReader, err := ingest.NewLedgerTransactionReader(ctx, backend, env.NetworkPassphrase, seq)
+		// txReader, err := ingest.NewLedgerTransactionReader(ctx, backend, publicPassword, seq)
+		changeReader, err := ingest.NewLedgerChangeReader(ctx, backend, env.NetworkPassphrase, seq)
 		if err != nil {
 			return []OperationTransformInput{}, err
 		}
+		txReader := changeReader.LedgerTransactionReader
 
 		for int64(len(opSlice)) < limit || limit < 0 {
 			tx, err := txReader.Read()
